fix: clean up and fix type and linting errors
